% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbGetQuery_safe}
\alias{dbGetQuery_safe}
\title{Execute arbitrary SQL code without forcing a data frame to be returned}
\usage{
dbGetQuery_safe(conn, statement, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{DBIConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{statement}{a character vector of length 1 containing SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Execute arbitrary SQL code without forcing a data frame to be returned
}

