% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbWipe}
\alias{dbWipe}
\title{Wipe out all tables in a database}
\usage{
dbWipe(conn, ...)
}
\arguments{
\item{conn}{a \link[DBI:DBIConnection-class]{DBI::DBIConnection} object}

\item{...}{arguments passed to \code{\link[DBI:dbRemoveTable]{DBI::dbRemoveTable()}}}
}
\description{
Wipe out all tables in a database
}
\details{
Finds all tables within a database and removes them
}
