% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_cleanup.R
\name{etl_cleanup}
\alias{etl_cleanup}
\alias{etl_cleanup.default}
\title{Cleanup after data loaded into DB}
\usage{
etl_cleanup(con, dir, ...)

\method{etl_cleanup}{default}(con, dir, ...)
}
\arguments{
\item{con}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{...}{arguments passed to methods}
}
\description{
Cleanup after data loaded into DB
}
\examples{
require(RPostgreSQL)
# connect directly
db <- etl_connect("mtcars", user = "postgres", password = "scem8467", host = "localhost", port = 5433)
etl_init(db)
etl_push(db)
etl_cleanup(db)
}
\seealso{
Other etl.functions: \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

