% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_load}
\alias{etl_load}
\alias{etl_load.mtcars}
\title{Load data into a database}
\usage{
etl_load(x, conn, ...)

\method{etl_load}{mtcars}(x, conn, ...)
}
\arguments{
\item{x}{an \link{etl} object.}

\item{conn}{A database connection (see )}

\item{...}{arguments passed on to methods}
}
\description{
Load data into a database
}

