% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_process.R
\name{etl_process}
\alias{etl_process}
\alias{etl_process.default}
\alias{etl_process.etl_mtcars}
\title{Process raw data into a format suitable for DB insert}
\usage{
etl_process(obj, ...)

\method{etl_process}{default}(obj, ...)

\method{etl_process}{etl_mtcars}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the \code{\link{etl}} object
}
\description{
Process raw data into a format suitable for DB insert
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", password = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_scrape() \%>\%
 etl_process() \%>\%
 str()
}
}
\seealso{
etl_init

Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}},
  \code{\link{etl_push.etl_mtcars}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}},
  \code{\link{etl_scrape.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

