% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_transform}
\alias{etl_transform}
\alias{etl_transform.default}
\alias{etl_transform.mtcars}
\title{Transform data}
\usage{
etl_transform(x, ...)

\method{etl_transform}{default}(x, ...)

\method{etl_transform}{mtcars}(x, ...)
}
\arguments{
\item{x}{an \link{etl} object.}

\item{...}{arguments passed on to methods}
}
\description{
Transform data
}

