% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_update.R
\name{etl_update}
\alias{etl_update}
\alias{etl_update.default}
\title{Update an existing DB}
\usage{
etl_update(con, ...)

\method{etl_update}{default}(con, ...)
}
\arguments{
\item{con}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{...}{arguments passed to methods}
}
\description{
Update an existing DB
}
\examples{
require(RPostgreSQL)
# connect directly
db <- etl_connect("mtcars", user = "postgres", password = "scem8467", host = "localhost", port = 5433)
etl_update(db)
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}}
}

