% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{files_that_dont_exist}
\alias{files_that_dont_exist}
\title{Check for files that don't already exist}
\usage{
files_that_dont_exist(src, dir)
}
\arguments{
\item{src}{a character vector of paths to filenames that you want to have}

\item{dir}{a path to a directory where  you want the files to be}
}
\value{
a logical vector of the same length as \code{src} indicating which
files in \code{src} do not already exist in \code{dir}.
}
\description{
Check for files that don't already exist
}

