% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_schema}
\alias{get_schema}
\title{Retrieve a pre-defined schema initialization script}
\usage{
get_schema(con, schema_name, pkg, ext = NULL, ...)
}
\arguments{
\item{con}{A database connection}

\item{schema_name}{The name of the schema}

\item{pkg}{The package defining the schema}

\item{ext}{The file extension used for the SQL schema file. If NULL (the default) it
be inferred from the \code{src_*} class of \code{con}. For example, if \code{con}
has class \code{\link[dplyr]{src_sqlite}} then \code{ext} will be \code{sqlite}.}

\item{...}{Currently ignored}
}
\description{
If the table definitions are at all non-trivial,
you may wish to include a pre-defined table schema. This function
will retrieve it.
}
\examples{

cars <- etl("mtcars")
get_schema(cars, "mtcars", "etl")
get_schema(cars, "my_crazy_schema", "etl")

}

