% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Func_Moda.R
\name{getmode}
\alias{getmode}
\title{Moda}
\usage{
getmode(v, na.rm = TRUE, zero.rm = TRUE)
}
\arguments{
\item{v}{Vetor para calculo}

\item{na.rm}{Booleno, tendo: TRUE -> remove valores nulo;
FALSE -> mantem valores nulos}

\item{zero.rm}{Booleno, tendo: TRUE -> remove valores zeros;
FALSE -> mantem valores zeros}
}
\value{
Retorna a moda
}
\description{
Realiza o calculo de moda de um vetor
}
