% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Func_Mapeamento_Campo.R
\name{map_vars}
\alias{map_vars}
\title{Mapeamento de variaveis}
\usage{
map_vars(
  data,
  mapping = list(Field = NULL, Type = NULL, Variable = NULL, Decimal = NULL),
  typeDate = "numeric"
)
}
\arguments{
\item{data}{Base de dados como dataframe}

\item{mapping}{Lista os campos utilizados no mapeamento, com suas
respectivas tipificacoes.
Sendo: Field: Campos do tibble a ser retornado;
       Type:  Tipo de campo ('Texto','Data",'Inteiro','Duplo);
       Variable: Nome do cabeçalho do dataframe;
       Decimal: Numero de casas decimais.
Ex: tribble(~Campo,       ~Tipo,      ~Variavel,      ~Decimal,
            "Campo1",     "Texto",    "CAMPO1_TBL",   NA,
            "Campo2",     "Data",     "CAMPO2_TBL",   NA,
            "Campo3",     "Inteiro",  "CAMPO3_TBL",   NA,
            "Campo4",     "Duplo",    "CAMPO4_TBL",    2)}

\item{typeDate}{Indica o formato que o campo com tificicacao DATA esta sendo
importada.
Sendo: 'numerical : numero;
       'dmy'      : dia-mes-ano;
       'dym'      : dia-ano-mes;
       'mdy'      : mes-dia-ano;
       'myd'      : mes-ano-dia;
       'ydm'      : ano-dia-mes;
       'ymd'      : ano-mes-dia;}
}
\value{
Retorna um tibble tipificado
}
\description{
Realiza o mapeamento e tipificacao de campos utilizados
                na composicao do relatorio
}
