% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Func_Importacao.R
\name{read}
\alias{read}
\title{Importacao de tabelas}
\usage{
read(
  file = list(Name = NULL, Dir = NULL, Extension = NULL, Sheet = NULL, Title = NULL,
    Seq = NULL, Id_Seq = NULL),
  BD = list(Conection = NULL, Query = NULL, Title = NULL)
)
}
\arguments{
\item{file}{Lista com parametros para inportacao dos arquivos.
Sendo: Name:   Nome do Arquivo;
       Dir:    Diretorio do Arquivo;
       Sheet:  Nome da Planilha caso seja xlsx
       Title:  Nome atribuido ao objeto importado
       Seq:    Sequencia de importacao, caso haja
               conjunto de arquivos a serem importados.
               Obs. O separador de sequencias é ';".
       Id_Seq: Id atribuido ao conjuto de dados das sequencias
Ex:tribble(~Name,      ~Dir,    ~Extension,   ~Sheet,   ~Title,   ~Seq,             ~Id_Seq
           Arquivo1,   Dir1,    xlsx,         Plan1,    Titulo1,  NA,               NA,
           Arquivo1,   Dir2,    csv,          NA,       Titulo2,  NA,               NA,
           ? - comum,  Dir3,    xlsx,         Plan1,    Titulo3,  2020.05;2020.06,  5;6)}

\item{BD}{Lista com os parametros para realizacao de consulta em
banco.
Sendo: Conection: Objeto com a conexao com o  banco;
       Query:     Consulta;
       Title:     Nome atribuido ao objeto.}
}
\value{
Retorna um dataframe
}
\description{
Realiza importacao de arquivo ou leitura de banco
}
