% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smart_download}
\alias{smart_download}
\title{Download only those files that don't already exist}
\usage{
smart_download(obj, src, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{src}{a character vector of paths to filenames that you want to have}

\item{...}{arguments passed to \code{\link[utils]{download.file}}}
}
\description{
Download only those files that don't already exist
}
\details{
Downloads only those files in \code{src} that are not already present
the \code{raw_dir} attribute of \code{obj}
files in \code{src} do not already exist in \code{dir}.
}

