% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_year_month}
\alias{valid_year_month}
\title{Ensure that years and month are within a certain time span}
\usage{
valid_year_month(years, months, begin = "1970-01-01", end = Sys.Date())
}
\arguments{
\item{years}{a vector of years}

\item{months}{a vector of months}

\item{begin}{the earliest valid date, defaults to the UNIX epoch}

\item{end}{the most recent valid date, defaults to today}
}
\description{
Ensure that years and month are within a certain time span
}
\details{
blah
}

