% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRunScript.R
\name{dbRunScript}
\alias{dbRunScript}
\title{Execute an SQL script}
\usage{
dbRunScript(con, script, ...)
}
\arguments{
\item{con}{a \code{DBIConnection-class} object}

\item{script}{path to an SQL script file}

\item{...}{currently ignored}
}
\value{
a list of results from \code{dbGetQuery} for each of the individual
SQL statements in \code{script}.
}
\description{
Execute an SQL script
}
\details{
The SQL script file must be \code{;} delimited.
}
\examples{
sql <- "SHOW TABLES; SELECT 1+1 as Two;"
tmpfile <- tempfile()
write(sql, file = tmpfile)

\dontrun{
if (require(RMySQL)) {
 con <- dbConnect(RMySQL::MySQL(), user = "r-user", password = "mypass", dbname = "mysql")
 dbRunScript(con, script = tmpfile)
 dbDisconnect(con)
}
}

}
\author{
Ben Baumer
}

