% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dbRunScript}
\alias{dbRunScript}
\title{Execute an arbitrary SQL script}
\usage{
dbRunScript(conn, script)
}
\arguments{
\item{conn}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{script}{Either a filename pointing to SQL script or
a character vector of length 1 containing SQL.}
}
\description{
Execute an arbitrary SQL script
}
\details{
The SQL script file must be \code{\"} delimited.
}

