% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRunScript.R
\name{dbRunScript}
\alias{dbRunScript}
\title{Execute an SQL script}
\usage{
dbRunScript(conn, script, ...)
}
\arguments{
\item{conn}{a \code{DBIConnection-class} object}

\item{script}{Either a filename pointing to SQL script or
a character vector of length 1 containing SQL.}

\item{...}{currently ignored}
}
\value{
a list of results from \code{dbGetQuery} for each of the individual
SQL statements in \code{script}.
}
\description{
Execute an SQL script
}
\details{
The SQL script file must be \code{;} delimited.
}
\examples{
sql <- "SHOW TABLES; SELECT 1+1 as Two;"
sql <- system.file("sql", "mtcars.mysql", package = "etl")

\dontrun{
if (require(RMySQL)) {
 con <- dbConnect(RMySQL::MySQL(), user = "r-user", password = "mypass", dbname = "mysql")
 dbRunScript(con, script = sql)
 dbDisconnect(con)
}
}

}

