% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbRunScript.R
\name{dbRunScript}
\alias{dbRunScript}
\title{Execute a SQL script}
\usage{
dbRunScript(con, script, ...)
}
\arguments{
\item{con}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{script}{path to a SQL script file}

\item{...}{arguments passed to methods}
}
\value{
a list of results from \code{\link[DBI]{dbGetQuery}}
}
\description{
Execute a SQL script
}
\details{
The SQL script file must be \code{\"} delimited.
}
\author{
Ben Baumer
}

