% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbWipe}
\alias{dbWipe}
\title{Wipe out all tables in a database}
\usage{
dbWipe(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Wipe out all tables in a database
}
\details{
Finds all tables within a database and removes them
}
