% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl.R
\name{etl}
\alias{etl}
\title{Initialize an etl object}
\usage{
etl(class)
}
\arguments{
\item{class}{String specifying the class of the etl object.
This class determines method dispatch of \code{etl_*()} functions.}
}
\value{
an \code{\link{etl}} object with a class specified by \code{class}.
}
\description{
Initialize an etl object
}
\examples{
\dontrun{
# define, extract, and transform a data source
e <- etl("mtcars")
e <- etl_extract(e)
e <- etl_transform(e)
str(e)

# create a database and load
library(dplyr)
db <- src_sqlite(tempfile(fileext = ".sqlite3"), create = TRUE)
etl_load(e, db)
tbl(db, "mtcars")
}
}

