% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_mtcars.R
\name{etl}
\alias{etl}
\title{etl-class}
\format{\describe{
 \item{pkg}{the name of the package corresponding to the data source}
 \item{con}{an objection of class \code{\link[DBI]{DBIConnection-class}}}
 \item{dir}{the directory where the raw data is stored}
 \item{init}{has the database been initialized?}
 \item{push}{was the push successful?}
}}
\usage{
etl()
}
\description{
etl-class
}

