% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_cleanup.R
\name{etl_cleanup}
\alias{etl_cleanup}
\alias{etl_cleanup.default}
\title{Cleanup after data loaded into DB}
\usage{
etl_cleanup(obj, ...)

\method{etl_cleanup}{default}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Cleanup after data loaded into DB
}
\examples{
require(RPostgreSQL)
# connect directly
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", password = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_create() \%>\%
 etl_cleanup() \%>\%
 str()
}
\seealso{
Other etl.functions: \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

