% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_connect.R
\name{etl_connect}
\alias{etl_connect}
\alias{etl_connect.default}
\title{Connect to a SQL database}
\usage{
etl_connect(x, db_con, dir = tempdir(), ...)

\method{etl_connect}{default}(x, db_con, dir = tempdir(), ...)
}
\arguments{
\item{x}{the name of the package that you wish to populate}

\item{db_con}{a \code{\link[dplyr]{src}} or
\code{\link[DBI]{DBIConnection-class}} object}

\item{dir}{a directory to store the raw data files}

\item{...}{arguments passed to methods}
}
\value{
an object of class \code{etl_*} and \code{\link{etl}}
}
\description{
Connect to a SQL database
}
\examples{
\dontrun{

# connect using dplyr
if (require(RPostgreSQL) & require(dplyr)) {
  db <- src_postgres("mtcars", user = "postgres",
  host = "localhost")
  cars <- etl_connect("mtcars", db)
  str(cars)
}
# connect using DBI
require(DBI)
con <- dbConnect(RPostgreSQL::PostgreSQL(), user = "postgres",
host = "localhost", dbname = "mtcars")
cars <- etl_connect("mtcars", db)
}
}
\seealso{
\code{\link{etl_create}}
}

