% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_connect.R
\name{etl_connect}
\alias{etl_connect}
\alias{etl_connect.default}
\title{Connect to a SQL database}
\usage{
etl_connect(x, db_con, dir = tempdir(), ...)

\method{etl_connect}{default}(x, db_con, dir = tempdir(), ...)
}
\arguments{
\item{x}{the name of the package that you wish to populate}

\item{db_con}{a \code{\link[dplyr]{src}} or
\code{\link[DBI]{DBIConnection-class}} object}

\item{dir}{a directory to store the raw data files}

\item{...}{arguments passed to methods}
}
\value{
an object of class \code{etl_*} and \code{\link{etl}}
}
\description{
Connect to a SQL database
}
\examples{
\dontrun{
require(RPostgreSQL)
# connect using dplyr
require(dplyr)
db <- src_postgres("mtcars", user = "postgres",
host = "localhost")
etl_cars <- etl_connect("mtcars", db)
str(etl_cars)

# connect using DBI
con <- dbConnect(RPostgreSQL::PostgreSQL(), user = "postgres",
host = "localhost", dbname = "mtcars")
etl_cars <- etl_connect("mtcars", db)
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}},
  \code{\link{etl_process.etl_mtcars}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}},
  \code{\link{etl_push.etl_mtcars}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}},
  \code{\link{etl_scrape.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

