% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_create.R
\name{etl_create}
\alias{etl_create}
\alias{etl_create.default}
\title{Create and populate a database from an online source}
\usage{
etl_create(con, ...)

\method{etl_create}{default}(con, ...)
}
\arguments{
\item{con}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{...}{arguments passed to methods}
}
\value{
the result of \code{\link{etl_update}}
}
\description{
Create and populate a database from an online source
}
\examples{
require(RPostgreSQL)
# connect directly
db <- etl_connect("mtcars", user = "postgres", password = "scem8467", host = "localhost", port = 5433)
etl_create(db)
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_init}}, \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

