% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_create.R
\name{etl_create}
\alias{etl_create}
\alias{etl_create.default}
\title{Create and populate a database from an online source}
\usage{
etl_create(obj, ...)

\method{etl_create}{default}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the result of \code{\link{etl_update}}
}
\description{
Create and populate a database from an online source
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
str(etl_create(etl_cars))
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_extract}},
  \code{\link{etl_extract.default}},
  \code{\link{etl_extract.etl_mtcars}};
  \code{\link{etl_init}}, \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_load}}, \code{\link{etl_load.default}},
  \code{\link{etl_load.etl_mtcars}};
  \code{\link{etl_transform}},
  \code{\link{etl_transform.default}},
  \code{\link{etl_transform.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

