% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract}
\alias{etl_extract}
\alias{etl_extract.default}
\alias{etl_extract.etl_mtcars}
\title{Scrape data from an online source}
\usage{
etl_extract(obj, ...)

\method{etl_extract}{default}(obj, ...)

\method{etl_extract}{etl_mtcars}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the (possibly modified) \code{\link{etl}} object
}
\description{
Scrape data from an online source
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_init() \%>\%
 etl_extract() \%>\%
 str()
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_load}}, \code{\link{etl_load.default}},
  \code{\link{etl_load.etl_mtcars}};
  \code{\link{etl_transform}},
  \code{\link{etl_transform.default}},
  \code{\link{etl_transform.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

