% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_extract}
\alias{etl_extract}
\alias{etl_extract.mtcars}
\title{Extract data and pull into R}
\usage{
etl_extract(obj, ...)

\method{etl_extract}{mtcars}(x, ...)
}
\arguments{
\item{...}{arguments passed onto methods}

\item{x}{an \link{etl} object.}
}
\description{
Extract data and pull into R
}

