% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_init.R
\name{etl_init}
\alias{etl_init}
\alias{etl_init.default}
\title{Initialize the DB with schema}
\usage{
etl_init(con, ...)

\method{etl_init}{default}(con, ...)
}
\arguments{
\item{con}{a \code{\link[DBI]{DBIConnection-class}} object}

\item{...}{arguments passed to methods}
}
\value{
the result from \code{\link{dbRunScript}}
}
\description{
Initialize the DB with schema
}
\examples{
require(RPostgreSQL)
# connect directly
db <- etl_connect("mtcars", user = "postgres", password = "scem8467", host = "localhost", port = 5433)
etl_init(db)
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

