% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_init.R
\name{etl_init}
\alias{etl_init}
\alias{etl_init.default}
\alias{etl_init.etl_mtcars}
\title{Initialize the DB with schema}
\usage{
etl_init(obj, ...)

\method{etl_init}{default}(obj, ...)

\method{etl_init}{etl_mtcars}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the \code{\link{etl}} object
}
\description{
Initialize the DB with schema
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
str(etl_cars)
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}};
  \code{\link{etl_extract}},
  \code{\link{etl_extract.default}},
  \code{\link{etl_extract.etl_mtcars}};
  \code{\link{etl_load}}, \code{\link{etl_load.default}},
  \code{\link{etl_load.etl_mtcars}};
  \code{\link{etl_transform}},
  \code{\link{etl_transform.default}},
  \code{\link{etl_transform.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

