% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_load.R
\name{etl_load}
\alias{etl_load}
\alias{etl_load.default}
\alias{etl_load.etl_mtcars}
\title{Push data to the DB}
\usage{
etl_load(obj, ...)

\method{etl_load}{default}(obj, ...)

\method{etl_load}{etl_mtcars}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the \code{\link{etl}} object
}
\description{
Push data to the DB
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_init() \%>\%
 etl_extract() \%>\%
 etl_transform() \%>\%
 etl_load()
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}};
  \code{\link{etl_extract}},
  \code{\link{etl_extract.default}},
  \code{\link{etl_extract.etl_mtcars}};
  \code{\link{etl_init}}, \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_transform}},
  \code{\link{etl_transform.default}},
  \code{\link{etl_transform.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

