% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_load}
\alias{etl_load}
\alias{etl_load.default}
\title{Create a database schema and populate it with data}
\usage{
etl_load(x, conn, schema, nms = names(x$data), ...)

\method{etl_load}{default}(x, conn, schema, nms = names(x$data), ...)
}
\arguments{
\item{x}{an \link{etl} object.}

\item{conn}{A database connection.}

\item{schema}{Either a filename pointing to SQL schema file or
a character vector containing the schema itself.}

\item{nms}{A character vector describing the names of table(s) to be populated.
This argument is ignored if schema is specified
(in that case, the table names in the schema are used).}

\item{...}{arguments passed on to methods}
}
\description{
Create a database schema and populate it with data
}

