% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_process.R
\name{etl_process}
\alias{etl_process}
\alias{etl_process.default}
\title{Process raw data into a format suitable for DB insert}
\usage{
etl_process(dir, ...)

\method{etl_process}{default}(dir, ...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\value{
a path to the directory
}
\description{
Process raw data into a format suitable for DB insert
}
\examples{
myDir <- "~/Desktop"
etl_scrape(dir = myDir)
etl_process()
}
\seealso{
etl_init

Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}}; \code{\link{etl_push}},
  \code{\link{etl_push.default}}; \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

