% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_push.R
\name{etl_push}
\alias{etl_push}
\alias{etl_push.default}
\alias{etl_push.etl_mtcars}
\title{Push data to the DB}
\usage{
etl_push(obj, ...)

\method{etl_push}{default}(obj, ...)

\method{etl_push}{etl_mtcars}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\value{
the \code{\link{etl}} object
}
\description{
Push data to the DB
}
\examples{
\dontrun{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_init() \%>\%
 etl_scrape() \%>\%
 etl_process() \%>\%
 etl_push() \%>\%
 str()
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}},
  \code{\link{etl_cleanup.etl_mtcars}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}},
  \code{\link{etl_init.etl_mtcars}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}},
  \code{\link{etl_process.etl_mtcars}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}},
  \code{\link{etl_scrape.etl_mtcars}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

