% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_scrape.R
\name{etl_scrape}
\alias{etl_scrape}
\alias{etl_scrape.default}
\title{Scrape data from an online source}
\usage{
etl_scrape(dir = tempdir(), ...)

\method{etl_scrape}{default}(dir = tempdir(), ...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\value{
a path to the directory
}
\description{
Scrape data from an online source
}
\examples{
etl_scrape(dir = "~/Desktop")
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_update}},
  \code{\link{etl_update.default}}
}

