% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_update.R
\name{etl_update}
\alias{etl_update}
\alias{etl_update.default}
\title{Update an existing DB}
\usage{
etl_update(obj, ...)

\method{etl_update}{default}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Update an existing DB
}
\examples{
require(RPostgreSQL)
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", password = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_init() \%>\%
 etl_update() \%>\%
 str()
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup}},
  \code{\link{etl_cleanup.default}};
  \code{\link{etl_connect}},
  \code{\link{etl_connect.default}};
  \code{\link{etl_create}},
  \code{\link{etl_create.default}}; \code{\link{etl_init}},
  \code{\link{etl_init.default}};
  \code{\link{etl_process}},
  \code{\link{etl_process.default}};
  \code{\link{etl_push}}, \code{\link{etl_push.default}};
  \code{\link{etl_scrape}},
  \code{\link{etl_scrape.default}}
}

